backend/requirements.txt
fastapi
uvicorn
PyJWT
bcrypt
psycopg2-binary